scraper/requirements.txt
beautifulsoup4
mysql-connector
urllib3
lxml
